third_party/blink/web_tests/external/wpt/css/css-values/animations/calc-interpolation-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <left> from [0px] to [calc(infinity * 1px)] at (-0.25) should be [-8.50705e+37px]
  assert_equals: expected "- 3.36e + 07px " but got "- 8.39e + 06px "
[FAIL] CSS Transitions: property <left> from [0px] to [calc(infinity * 1px)] at (0.25) should be [8.50705e+37px]
  assert_equals: expected "3.36e + 07px " but got "8.39e + 06px "
[FAIL] CSS Transitions: property <left> from [0px] to [calc(infinity * 1px)] at (0.5) should be [1.70141e+38px]
  assert_equals: expected "3.36e + 07px " but got "1.68e + 07px "
[FAIL] CSS Transitions: property <left> from [0px] to [calc(infinity * 1px)] at (0.75) should be [2.552115e+38px]
  assert_equals: expected "3.36e + 07px " but got "2.52e + 07px "
[FAIL] CSS Transitions with transition: all: property <left> from [0px] to [calc(infinity * 1px)] at (-0.25) should be [-8.50705e+37px]
  assert_equals: expected "- 3.36e + 07px " but got "- 8.39e + 06px "
[FAIL] CSS Transitions with transition: all: property <left> from [0px] to [calc(infinity * 1px)] at (0.25) should be [8.50705e+37px]
  assert_equals: expected "3.36e + 07px " but got "8.39e + 06px "
[FAIL] CSS Transitions with transition: all: property <left> from [0px] to [calc(infinity * 1px)] at (0.5) should be [1.70141e+38px]
  assert_equals: expected "3.36e + 07px " but got "1.68e + 07px "
[FAIL] CSS Transitions with transition: all: property <left> from [0px] to [calc(infinity * 1px)] at (0.75) should be [2.552115e+38px]
  assert_equals: expected "3.36e + 07px " but got "2.52e + 07px "
Harness: the test ran to completion.
